Reorganizing ide/lib.rs
