+++ src/rust/ide/src/config.rs
@@ -1,4 +1,21 @@
 //! This module provides constants used throughout the crate.
 
-/// Visualization folder where IDE can look for user-defined visualizations per project.
-pub const PROJECT_VISUALIZATION_FOLDER:&str = "visualization";
+use crate::constants;
+
+/// Configuration data necessary to initialize IDE.
+///
+/// Eventually we expect it to be passed to IDE from an external source.
+#[derive(Clone,Debug)]
+pub struct SetupConfig {
+    /// WebSocket endpoint of the project manager service.
+    pub project_manager_endpoint:String
+}
+
+impl SetupConfig {
+    /// Provisional initial configuration that can be used during local deployments.
+    pub fn new_local() -> SetupConfig {
+        SetupConfig {
+            project_manager_endpoint:constants::PROJECT_MANAGER_ENDPOINT.into()
+        }
+    }
+}
